Update listview refresh interval to 500 milliseconds
With threading, this is less of a concern and can be updated quickly which
may also mean more often.  Also, expose LVDelay as a config file option
to tune further, though "undocumented".  This must be manually set in the
config file and cannot be changed during runtime from Settings dialog.