Make Escape key smarter based on docking or floating dialog
If the dialog is just opened as a floating window, Escape key should - and
currently does - close the floating dialog window. If however the dialog
is docked to the Notepad++ window in one of the four (top, bottom, left,
or right) docking spots, Escape key should not close the dialog; rather,
it should set focus to the current Scintilla editing component. The
thought is if the dialog is docked, the user may want to use it multiple
times while swapping back and forth to navigate through the document.